Show ghost cells instead of the no results
